Sửa bảng + sửa code của SampleGroup: OpenTime, CloseTime -> kiểu int: số giây tính từ 0 giờ trong ngày

+++ eLTMS/eLTMS.Models/Models/dto/SampleDto.cs
@@ -16,9 +16,9 @@ namespace eLTMS.Models.Models.dto
 
         public int SampleDuration { get; set; }
 
-        public double OpenTime { get; set; }
+        public int OpenTime { get; set; }
 
-        public double CloseTime { get; set; }
+        public int CloseTime { get; set; }
 
     }
 }
